dont handle fileLimit error on streams
